requirements.txt
google-adk>=0.1.0
google-generativeai>=0.3.0